unifies dataset transforms & signal transforms; now there are only "transforms" can be applied to both
